+++ CHANGELOG.md
@@ -37,6 +37,40 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+## State Machine Breaking
+
+* (intrarelayer) [\#119](https://github.com/tharsis/evmos/issues/119) Register `x/intrarelayer` proposal types on governance module.
+
+### Improvements
+
+* (app) [\#128](https://github.com/tharsis/evmos/pull/128) Add ibc-go `TestingApp` interface.
+* (ci) [\#117](https://github.com/tharsis/evmos/pull/117) Enable automatic backport of PRs.
+* (deps) [\#135](https://github.com/tharsis/evmos/pull/135) Bump Ethermint version to [`v0.9.0`](https://github.com/tharsis/ethermint/releases/tag/v0.9.0)
+* (ci) [\#136](https://github.com/tharsis/evmos/pull/136) Deploy `evmos` docker container to [docker hub](https://hub.docker.com/u/tharsishq) for every versioned releases
+
+### Bug Fixes
+
+* (build) [\#116](https://github.com/tharsis/evmos/pull/116) Fix `build-docker` command
+
+## [v0.3.0] - 2021-11-24
+
+### API Breaking
+
+* (intrarelayer) [\#99](https://github.com/tharsis/evmos/pull/99) Rename `enable_e_v_m_hook` json parameter to `enable_evm_hook`.
+
+### Improvements
+
+* (deps) [\#110](https://github.com/tharsis/evmos/pull/110) Bump Ethermint version to [`v0.8.1`](https://github.com/tharsis/ethermint/releases/tag/v0.8.1)
+* (intrarelayer) [\#107](https://github.com/tharsis/evmos/pull/107) Add IBC validation
+* (cmd) [\#105](https://github.com/tharsis/evmos/pull/105) Improve testnet command to include JSON-RPC client.
+
+## Bug Fixes
+
+* (intrarelayer) [\#109](https://github.com/tharsis/evmos/pull/109) Fix hardcoded intrarelayer nonce and `UpdateTokenPairERC20` proposal to support ERC20s with 0 decimals.
+* (intrarelayer) [\#102](https://github.com/tharsis/evmos/pull/102) Add `convert-erc20` cmd
+
+## [v0.2.0] - 2021-11-17
+
 ### Features
 
 * (intrarelayer) [\#82](https://github.com/tharsis/evmos/pull/82) Intrarelayer module
@@ -53,7 +87,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 * (deps) [\#64](https://github.com/tharsis/evmos/pull/64) Bump Ethermint version to `v0.7.2`
 
-## Bug Fixes
+### Bug Fixes
 
 * (cmd) [\#41](https://github.com/tharsis/evmos/pull/41) Fix `debug` command.
 
@@ -65,13 +99,13 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [v0.1.1] - 2021-10-07
 
-## Bug Fixes
+### Bug Fixes
 
 * (build) [\#30](https://github.com/tharsis/evmos/pull/30) Fix `version` command.
 
 ## [v0.1.0] - 2021-10-07
 
-## Improvements
+### Improvements
 
 * (cmd) [\#26](https://github.com/tharsis/evmos/pull/26) Use config on genesis accounts.
 * (deps) [\#28](https://github.com/tharsis/evmos/pull/28) Bump Ethermint version to `v0.7.0`